+++ src/main/java/com/prgms/management/customer/exception/CustomerException.java
@@ -1,9 +1,7 @@
 package com.prgms.management.customer.exception;
 
-public class CustomerException extends Exception{
-    private static final long serialVersionUID = 7544775631058799907L;
-
-    protected CustomerException(String message) {
+public class CustomerException extends RuntimeException {
+    public CustomerException(String message) {
         super(message);
     }
 }
