+++ src/main/java/com/prgms/management/customer/model/Customer.java
@@ -1,4 +1,4 @@
-package com.prgms.management.customer.entity;
+package com.prgms.management.customer.model;
 
 import java.util.UUID;
