+++ src/main/java/com/prgms/management/customer/model/CustomerType.java
@@ -1,4 +1,4 @@
-package com.prgms.management.customer.entity;
+package com.prgms.management.customer.model;
 
 public enum CustomerType {
     WHITE, BLACK
