+++ src/main/java/com/prgms/management/customer/repository/BlackCustomerRepository.java
@@ -1,11 +1,9 @@
 package com.prgms.management.customer.repository;
 
-import com.prgms.management.customer.entity.Customer;
-import com.prgms.management.customer.exception.CustomerException;
+import com.prgms.management.customer.model.Customer;
 
 import java.util.List;
 
 public interface BlackCustomerRepository {
-    List<Customer> findAll() throws CustomerException;
-
+    List<Customer> findAll();
 }
