+++ src/main/java/com/prgms/management/customer/repository/FileBlackCustomerRepository.java
@@ -1,15 +1,16 @@
 package com.prgms.management.customer.repository;
 
-import com.prgms.management.customer.entity.Customer;
-import com.prgms.management.customer.entity.CustomerType;
 import com.prgms.management.customer.exception.CustomerException;
-import com.prgms.management.customer.exception.CustomerListEmptyException;
+import com.prgms.management.customer.model.Customer;
+import com.prgms.management.customer.model.CustomerType;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.DefaultResourceLoader;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Repository;
 
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -24,18 +25,19 @@ public class FileBlackCustomerRepository implements BlackCustomerRepository {
     }
 
     @Override
-    public List<Customer> findAll() throws CustomerException {
+    public List<Customer> findAll() {
         List<Customer> customers = new ArrayList<>();
 
         try (BufferedReader bufferedReader = new BufferedReader(new FileReader(resource.getFile()))) {
-            String item;
-            while ((item = bufferedReader.readLine()) != null) {
-                String[] array = item.split(",");
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                String[] array = line.split(",");
                 customers.add(new Customer(CustomerType.BLACK, UUID.fromString(array[0]), array[1]));
             }
         } catch (IOException e) {
-            throw new CustomerListEmptyException();
+            throw new CustomerException(e.getMessage());
         }
+
         return customers;
     }
 }
